+++ lib/l10n/intl_en.arb
@@ -426,7 +426,6 @@
   "randomSurahSelection": "Random Surah Selection",
   "selectSurah": "Select a Surah",
   "initializingAutoReading": "Initializing in progress...",
-  "initializingAutoReading": "Initializing in progress...",
   "prayerTimeNotification": "{salahName} time ({prayerTime}) notification",
   "@prayerTimeNotification": {
     "description": "Notifcation that will be displayed when app is backgrounded.",
